Removes NavBarView, moves hiding logic into hook
