+++ src/components/Header/Nav/index.ts
@@ -1,3 +1,2 @@
 export { NavBar } from "./NavBar"
-export { NavBarView } from "./NavBarView"
 export type { devHubCategories } from "./config"
